dashboard/requirements.txt
django 
channels
channels-redis
markdown 
requests
grpcio
grpcio-tools 
djangorestframework 
django-ratelimit 
django-redis
redit
daphne
